:truck: Move hint to its own directory
