+++ Cargo.toml
@@ -1,26 +1,27 @@
 [package]
 description = "Run efficiently your unit tests written in cairo"
-edition     = "2021"
-name        = "cairo-foundry"
-version     = "0.1.0"
+edition = "2021"
+name = "cairo-foundry"
+version = "0.1.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cairo-rs    = { git = "https://github.com/tdelabro/cairo-rs", branch = "cairo-foundry" }
-clap        = { version = "3.2.6", features = ["derive"], default-features = false }
-colored     = "2.0.0"
-dirs        = "4.0"
-env_logger  = "0.9.0"
+cairo-rs = {git = "https://github.com/tdelabro/cairo-rs", branch = "cairo-foundry"}
+clap = {version = "3.2.6", features = ["derive"], default-features = false}
+colored = "2.0.0"
+dirs = "4.0"
+env_logger = "0.9.0"
 lazy_static = "1.4.0"
-log         = "0.4.17"
-num-bigint  = "0.4.3"
-num-traits  = "0.2.15"
-rayon       = "1.5.3"
-regex       = { version = "1.6.0", default-features = true }
-serde       = { version = "1.0", features = ["derive"] }
-serde_json  = "1.0"
-thiserror   = "1.0.37"
-uuid        = { version = "1.2.1", features = ["v4"] }
-walkdir     = "2.3.2"
-which       = "4.3.0"
+log = "0.4.17"
+num-bigint = "0.4.3"
+num-traits = "0.2.15"
+rayon = "1.5.3"
+regex = {version = "1.6.0", default-features = true}
+rstest = "0.15.0"
+serde = {version = "1.0", features = ["derive"]}
+serde_json = "1.0"
+thiserror = "1.0.37"
+uuid = {version = "1.2.1", features = ["v4"]}
+walkdir = "2.3.2"
+which = "4.3.0"
