+++ src/hints/mod.rs
@@ -15,6 +15,9 @@ use uuid::Uuid;
 
 pub const EXECUTION_UUID_VAR_NAME: &str = "cairo-foundry-execution-uuid";
 
+mod assert_revert;
+pub use assert_revert::assert_revert;
+
 lazy_static! {
 	static ref HINT_OUTPUT_BUFFER: RwLock<HashMap<Uuid, String>> = RwLock::new(HashMap::new());
 }
@@ -63,21 +66,4 @@ pub fn skip(
 	_constants: &HashMap<String, BigInt>,
 ) -> Result<(), VirtualMachineError> {
 	Err(VirtualMachineError::CustomHint("skip".to_string()))
-}
-
-pub fn assert_revert(
-	vm: &mut VirtualMachine,
-	_exec_scopes: &mut ExecutionScopes,
-	_ids_data: &HashMap<String, HintReference>,
-	_ap_tracking: &ApTracking,
-	_constants: &HashMap<String, BigInt>,
-) -> Result<(), VirtualMachineError> {
-	match vm.step_instruction() {
-		Ok(_) => Err(VirtualMachineError::CustomHint(
-			"assert_revert_did_not_revert".to_string(),
-		)),
-		Err(_) => Err(VirtualMachineError::CustomHint(
-			"assert_revert_reverted".to_string(),
-		)),
-	}
 }
